Move PlexGuid to plextraktsync.plex.guid namespace
